Extract out PersistentStorage into a separate class
